Refactor item and digitalobject repo
