+++ dc-cudami-server/dc-cudami-server-backend-api/src/main/java/de/digitalcollections/cudami/server/backend/api/repository/identifiable/entity/work/ItemRepository.java
@@ -2,7 +2,6 @@ package de.digitalcollections.cudami.server.backend.api.repository.identifiable.
 
 import de.digitalcollections.cudami.server.backend.api.repository.exceptions.RepositoryException;
 import de.digitalcollections.cudami.server.backend.api.repository.identifiable.entity.EntityRepository;
-import de.digitalcollections.model.identifiable.entity.digitalobject.DigitalObject;
 import de.digitalcollections.model.identifiable.entity.item.Item;
 import de.digitalcollections.model.identifiable.entity.manifestation.Manifestation;
 import de.digitalcollections.model.list.paging.PageRequest;
@@ -13,17 +12,6 @@ import java.util.UUID;
 
 /** Repository for Item persistence handling. */
 public interface ItemRepository extends EntityRepository<Item> {
-
-  default PageResponse<DigitalObject> findDigitalObjects(Item item, PageRequest pageRequest)
-      throws RepositoryException {
-    if (item == null) {
-      throw new IllegalArgumentException("find failed: given object must not be null");
-    }
-    return findDigitalObjects(item.getUuid(), pageRequest);
-  }
-
-  PageResponse<DigitalObject> findDigitalObjects(UUID itemUuid, PageRequest pageRequest)
-      throws RepositoryException;
 
   default PageResponse<Item> findItemsByManifestation(
       Manifestation manifestation, PageRequest pageRequest) throws RepositoryException {
